src/requirements.txt
discord.py==2.3.2
openai==1.3.7
paramiko==3.3.1
tiktoken==0.5.1
